+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ### vNEXT
 - Modified `ign-right-drawer` to use `mat-drawer`, to unfocus/darken background content, to close on backdrop click and escape key press, and to make content scrollable
 
+### v0.8.0
+- Added an aria-describedby between the button and the editable field within the `ign-editable-field` component
+- Added an aria-label to the button within the `ign-editable-field` component
+
 ### v0.7.0
 - Deprecated ign-nav-list, it is recommended to use mat-nav-list instead
 - Deprecated ign-divider, it is recommended to use mat-divider instead